+++ src/main/ts/ephox/alloy/keying/FlowType.ts
@@ -87,4 +87,4 @@ const getKeyupRules: () => Array<KeyRules.KeyRule<FlowConfig, Stateless>> = Fun.
   KeyRules.rule(KeyMatch.inSet(Keys.SPACE()), KeyingTypes.stopEventForFirefox)
 ])
 
-export default KeyingType.typical(schema, NoState.init, getKeydownRules, getKeyupRules, Option.some(focusIn));
+export default KeyingType.typical(schema, NoState.init, getKeydownRules, getKeyupRules, () => Option.some(focusIn));