+++ src/main/ts/ephox/alloy/keying/KeyingModeTypes.ts
@@ -4,19 +4,29 @@ import { Element } from '@ephox/sugar';
 import { AlloyComponent } from '../api/component/ComponentApi';
 import { FocusManager } from '../api/focus/FocusManagers';
 import { BehaviourState, Stateless } from '../behaviour/common/BehaviourState';
-import { NativeSimulatedEvent } from '../events/SimulatedEvent';
+import { NativeSimulatedEvent, SimulatedEvent } from '../events/SimulatedEvent';
 
 // TODO: Fix this.
 export type KeyHandlerApi<C, S> = (comp: AlloyComponent, se: NativeSimulatedEvent, config?: C, state?: S) => Option<boolean>;
 
 export type KeyRuleHandler<C, S> = (comp: AlloyComponent, se: NativeSimulatedEvent, config: C, state?: S) => Option<boolean>;
 
+
+export enum FocusInsideModes {
+  OnFocusMode ='onFocus',
+  OnEnterOrSpaceMode = 'onEnterOrSpace',
+  OnApiMode = 'onApi'
+}
+
 export interface GeneralKeyingConfigSpec {
   focusManager?: FocusManager;
+  focusInside?: FocusInsideModes;
 }
 
 export interface GeneralKeyingConfig {
   focusManager: FocusManager;
+  sendFocusIn: <C extends GeneralKeyingConfig, S>(conf: C) => Option<(comp: AlloyComponent, config: C, state: S, evt?: SimulatedEvent<any>) => void>;
+  focusInside: FocusInsideModes;
 }
 
 export interface TabbingConfigSpec<C extends TabbingConfig> extends GeneralKeyingConfigSpec {